questions/bipartite_graphs.py
"""
Problem: Bipartite Graph

Description:
Given an undirected graph represented by a 2D array where graph[u] lists nodes adjacent to node u, determine if the graph is bipartite. A graph is bipartite if the nodes can be partitioned into two independent sets A and B such that every edge in the graph connects a node in set A and a node in set B.

Function Signature:
def isBipartite(graph: List[List[int]]) -> bool:

Inputs:
    - graph (List[List[int]]): A 2D array representing connections between nodes.

Returns:
    - bool: True if the graph is bipartite, otherwise False.

Examples:
1. Input: graph = [[1,2,3],[0,2],[0,1,3],[0,2]]
   Output: false
   Explanation: There is no possible partition that satisfies the bipartite conditions.

2. Input: graph = [[1,3],[0,2],[1,3],[0,2]]
   Output: true
   Explanation: The graph can be partitioned into two sets: {0, 2} and {1, 3}.

Constraints:
    - graph.length == n
    - 1 <= n <= 100
    - 0 <= graph[u].length < n
    - 0 <= graph[u][i] <= n - 1
    - graph[u] does not contain u.
    - All values in graph[u] are unique.
    - If graph[u] contains v, then graph[v] contains u.

Notes:
    - A Depth First Search (DFS) or Breadth First Search (BFS) can be employed to traverse through the graph and check if it's bipartite.
    - A coloring approach can be used, where adjacent nodes must have different colors.
    - If during traversal we find two adjacent nodes with the same color, the graph is not bipartite.

Tags:
    - Graph
    - Depth First Search
    - Breadth First Search
"""

from typing import List

def isBipartite(graph: List[List[int]]) -> bool:
  return False